Config fixes
