+++ cypress/support/component.tsx
@@ -1,7 +1,7 @@
 import { mount } from "cypress/react18";
 import "cypress-real-events";
-import AppContextProvider from "@/app";
+import AppSetup from "@/app";
 
 Cypress.Commands.add("mount", (component, options) => {
-    return mount(<AppContextProvider>{component}</AppContextProvider>, options);
+    return mount(<AppSetup>{component}</AppSetup>, options);
 });
